\name{fit or miss-class}
\Rdversion{1.1}
\docType{class}
\alias{fit or miss-class}
\alias{number or miss-class}

\title{Class Unions}
\description{
  The implemetation of the \code{Reaper} class makes use of two class
  unions in order to deal with cases when we are unable to fit a
  model. The \code{number or miss} class union can either hold a
  numeric vector or a logical vector; the \code{fit or miss} class union
  can hold either a fitted \code{movMF} object or a logical vector.
}
\section{Objects from the Class}{Virtual Class: No objects may be created from it.}
\section{Methods}{
No methods are defined with class "number or miss" or  "fit or miss" in the signature.
}
\author{
Kevin R. Coombes ,krc@silicovore.com>
}
\keyword{classes}
