\name{getSplit-methods}
\docType{methods}
\alias{getSplit-methods}
\alias{getSplit,Reaper-method}
\alias{getSplit,Thresher-method}
\alias{getColors-methods}
\alias{getColors,Reaper-method}
\alias{getColors,Thresher-method}
\title{Color Palettes for Threshing and Reaping}
\description{
Both the \code{Thresher} an \code{Reaper} classes define default color
schemes for the primary objects being clustered and for the features
being used to cluster them. These color schemes are used to produce
color bars on heatmaps produce using two-way clustering.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Reaper")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Thresher")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
